requirements.txt
Flask==2.0.1
pytorch==1.9.0
transformers==4.10.0
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3
scikit-learn==0.24.2
jupyter==1.0.0
pytest==6.2.5